feat(notifications): merge read/dismiss action, un-nest account menu list
- Replace separate 'Alle gelesen' + 'Alle erledigt' buttons with a single 'Alle erledigt' (dismiss-all) action; drop the now-dead useMarkAllAsRead hook and markAllNotificationsAsRead client (backend endpoint kept for mobile).
- Render the list inline in the account dropdown above Einstellungen instead of in a nested bordered card, with bounding separators.
- Make the list scroll natively (max-h + overflow-y-auto) so menu items below stay reachable.
- Extract grouping/formatting helpers into utils/notificationGrouping.ts.
